Eliminate ColumnNameGenerator
Handling of column names is being moved into the appropriate tables.
This change was made because it is really each individual table's
implementation that dictates how column names should be generated. For
example: SQL requires a delicate dance with an RDBMS. But a simple text
file might need auto-generated names because it lacks even basic
headers. Rather than trying to formalize this, each implementation must
simply adhere to the Metadata trait.