Eliminate SqlQueryPackage
Functionality in SqlQueryPackage has now been separated into the
SqlMetadata trait as well as SqlRelation. As mentioned in other
commits, the primary driver for this is the need to centralize the
iteration over JDBC ResultSets and also to try as hard as humanly
possible not to execute multiple simultaneous database queries. All
metadata from the Metadata trait is marked as lazy to ensure that
nothing is executed until the last possible moment. This also
eliminates some duplicated state and metadata held in the SqlIterator
and query packages.