Remove SQLRelationColumn
The planned move to make transformations a core part of the API makes
SqlRelationColumn unecessary. A column of data from a SQL relation is
no different than a simple transform that throws out all the columns
but one. Single columns of data, while convenient are largely a nice to
have, not a major feature that rises to the level of needing support
from a specific object.