Add limited support for transfomed sql tables
In order to simplify the ETL process and avoid the end-user from having
to manage lots of state inside ETL scripts, there is now a
SqlTransform. The SqlTransform wraps another DataTable and applies a
function which transforms the rows of the table. The result is a new
DataTable which may columns that differ in number and type from the
source table.