Fix bug where connection properties are ignored for all databases
The Java Properties constructor optionally takes another Properties
object. However, this merely creates an **empty** properties object.
That is, the keys are there, but the values are not. The Properties
object provides no good facility for creating a new "clone" of an
existing object. While not ideal, we can get around this because
Properties (oddly) subclasses HashMap. So a "putAll" does the trick.