Add explicit dbname to mysql properties only for testing purposes
MySQL requires the schema to be in the URL and not a property, so
there's not a good way to read this info in. The tests depend on the
schema, but this may be removed later if it's not needed (I suspect it
isn't now that DataExpress handles schemas under the hood).