Fix error where tests were looking at the data source instead of target
It's unclear why this bug was exposed in Travis-CI and not locally.
Essentially the qurey_and_count method was checking the source database
for changes (which remains unchanged) and so it was never seeing the
updates. It's unclear why this works locally. My only suspicion is that
in Travis-CI we have different users and permissions and these revealed
the issue.