Remove columnNameGenerator and add companion object to MagicMarshaller
With the reitiring of columnNameGenerator, the lazy nature of column
name gneration needs to be preserved. Unfortunately, case classes can't
have lazy parameters owing to the way they work. To work around this,
an ordinary class with a companion object was created to at least
emulate case class usage. If this is insufficient, future refactoring
may be needed.