Update Metadata to use functions instead of vals for columnNames, dataTypes, and columnCount
In order to preserve the lazy evaluation of file-based operations, the
Metadata trait was changed to implement the various elements as
functions instead of vals. While all tests now pass, the need for this
is a bit dubious. In general, DataTables should be immutable, and
therefore, lazy vals should be sufficient. If something like column
names changes as a result of an operation, those changes should be
reflected in a brand new DataTable. This is a more functional way to go
and avoids strange state issues with order of operations.